Fix two problems with the Art Appreciation support
Since ShowAppreciation is static, we can't assume we have the art appreciation type. As it might not be loaded in yet.
!zoom was using an int which isn't a float so it wasn't waiting since I changed the Zoom method to expect an float.

+++ TwitchPlaysAssembly/Src/Commands/ModuleCommands.cs
@@ -424,7 +424,7 @@ static class ModuleCommands
 			if ((!timed || !command) && (zooming || tilt))
 			{
 				MusicPlayer musicPlayer = null;
-				int delay = 2;
+				float delay = 2;
 				if (timed)
 				{
 					delay = groups["time"].Value.TryParseInt() ?? groups["stime"].Value.TryParseInt() ?? 2;
